intermediate for amount_t

+++ src/xbridge/currency.h
@@ -102,7 +102,7 @@ namespace ccy {
     class Asset {
         // variables
         Currency _currency{};
-        Amount _amount{0};
+        Amount _amount{uint64_t(0)};
 
     public:
